+++ app/src/main/java/com/mskd/flux/screens/player/controllers/PlayerStateHolder.kt
@@ -39,6 +39,7 @@ import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
 import java.util.Locale
+import kotlin.math.roundToInt
 
 
 @OptIn(UnstableApi::class)
@@ -284,6 +285,11 @@ class PlayerStateHolder(
 
     fun updateProgress(progress: Long) {
         player.seekTo(progress)
+    }
+
+    fun changeVolume(delta: Float) : Int {
+        player.volume = (player.volume + delta).coerceIn(0f, 1f)
+        return (player.volume * 100).roundToInt()
     }
 
     fun playMedia(media: Media?) {