+++ app/src/main/java/com/mskd/flux/utils/enums/Side.kt
@@ -1,0 +1,5 @@
+package com.mskd.flux.utils.enums
+
+enum class Side {
+    LEFT, RIGHT
+}