+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## Version 1.X.X
+## Version 1.2.0
 
+- Free orientation within the player
+- Adjust volume and brightness by swipe
+- Add full screen mode in landscape
 - Update dependencies